Added 'Recalculate Scores', changed rating scale.
There is now a link in the upper right corner that allows you to
recalculate all the venue scores. A venue's score is not the sum of all
votes, as it was previously. It is, instead, the average of all the
existing ratings for that venue.

The user-applied ratings are now 0, 2, 4, 6, 8, 10, or X. The X value
means "no rating at all", which means that the user contributes nothing
to the sum or count used in computing the venue's average.